// server/src/models/enums.ts
export enum UserRole {
  REQUESTOR = 'REQUESTOR',
  HEAD_OF_DEPARTMENT = 'HEAD_OF_DEPARTMENT',
  HEAD_HR_ADMIN = 'HEAD_HR_ADMIN',
  CHIEF_EXECUTIVE_OFFICER = 'CHIEF_EXECUTIVE_OFFICER',
  CHAIRMAN = 'CHAIRMAN',
}

export enum ApprovalStatus {
  PENDING = 'PENDING',
  APPROVED = 'APPROVED',
  REJECTED = 'REJECTED',
}

export enum RequestStatus {
  DRAFT = 'DRAFT',
  SUBMITTED = 'SUBMITTED',
  HOD_APPROVED = 'HOD_APPROVED',
  HOD_REJECTED = 'HOD_REJECTED',
  HR_APPROVED = 'HR_APPROVED',
  HR_REJECTED = 'HR_REJECTED',
  CEO_APPROVED = 'CEO_APPROVED',
  CEO_REJECTED = 'CEO_REJECTED',
  CHAIRMAN_APPROVED = 'CHAIRMAN_APPROVED',
  CHAIRMAN_REJECTED = 'CHAIRMAN_REJECTED',
  COMPLETED = 'COMPLETED',
}